Show total amount

+++ WorldOfPAYBACK/Models/Codables/Value.swift
@@ -2,7 +2,7 @@
 import Foundation
 
 struct Value: Codable {
-	let amount: Int
+	let amount: Decimal // Using Decimal instead of Int, as amount can be in decimals
 	let currency: String
 }
